@@ -28,48 +30,48 @@
     "@angular/platform-browser-dynamic": "16.2.4",
     "@angular/platform-server": "16.2.4",
     "@angular/router": "16.2.4",
-    "@delon/abc": "^15.2.1",
-    "@delon/acl": "^15.2.1",
-    "@delon/auth": "^15.2.1",
-    "@delon/cache": "^15.2.1",
-    "@delon/chart": "^15.2.1",
-    "@delon/form": "^15.2.1",
-    "@delon/mock": "^15.2.1",
-    "@delon/theme": "^15.2.1",
-    "@delon/util": "^15.2.1",
-    "@ngx-translate/core": "^13.0.0",
-    "@ngx-translate/http-loader": "^6.0.0",
-    "ajv": "^8.10.0",
+    "@delon/abc": "^16.3.0",
+    "@delon/acl": "^16.3.0",
+    "@delon/auth": "^16.3.0",
+    "@delon/cache": "^16.3.0",
+    "@delon/chart": "^16.3.0",
+    "@delon/form": "^16.3.0",
+    "@delon/mock": "^16.3.0",
+    "@delon/theme": "^16.3.0",
+    "@delon/util": "^16.3.0",
     "aspnet-prerendering": "^3.0.1",
     "bootstrap": "^4.4.1",
-    "core-js": "^3.3.3",
     "date-fns": "^2.12.0",
-    "jquery": "3.5.0",
     "ng-zorro-antd": "^16.1.0",
-    "ngx-tinymce": "^13.0.0",
-    "ngx-ueditor": "^13.0.0",
-    "oidc-client": "^1.9.1",
-    "popper.js": "^1.16.0",
     "rxjs": "~7.8.0",
     "screenfull": "^5.0.2",
     "tslib": "^2.0.0",
     "wangeditor": "^4.7.4",
-    "zone.js": "~0.13.0",
-    "ajv-formats": "^2.1.1"
+    "zone.js": "~0.13.0"
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "^16.2.1",
+    "@angular-eslint/builder": "~16.1.1",
+    "@angular-eslint/eslint-plugin": "~16.1.1",
+    "@angular-eslint/eslint-plugin-template": "~16.1.1",
+    "@angular-eslint/schematics": "~16.1.1",
+    "@angular-eslint/template-parser": "~16.1.1",
     "@angular/cli": "^16.2.1",
     "@angular/compiler-cli": "16.2.4",
-    "@angular/language-service": "^16.2.4",
-    "@delon/testing": "^15.2.1",
+    "@angular/language-service": "^16.2.0",
+    "@delon/testing": "^16.3.0",
     "@types/jasmine": "~3.6.0",
     "@types/jasminewd2": "~2.0.8",
-    "@types/jszip": "^3.1.7",
-    "@types/node": "^12.11.1",
-    "antd-theme-generator": "^1.1.9",
-    "codelyzer": "^6.0.0",
-    "husky": "^7.0.4",
+    "@typescript-eslint/eslint-plugin": "~6.4.1",
+    "@typescript-eslint/parser": "~6.4.1",
+    "eslint": "^8.48.0",
+    "eslint-config-prettier": "~9.0.0",
+    "eslint-plugin-deprecation": "~1.5.0",
+    "eslint-plugin-import": "~2.28.1",
+    "eslint-plugin-jsdoc": "~46.5.0",
+    "eslint-plugin-prefer-arrow": "~1.2.3",
+    "eslint-plugin-prettier": "~5.0.0",
+    "husky": "^8.0.3",
     "jasmine-core": "~3.8.0",
     "jasmine-spec-reporter": "~5.0.0",
     "karma": "~6.3.16",
