@@ -1,4 +1,3 @@
-// 请参考：https://ng-alain.com/docs/i18n
 import { Platform } from '@angular/cdk/platform';
 import { registerLocaleData } from '@angular/common';
 import ngEn from '@angular/common/locales/en';
@@ -6,21 +5,21 @@ import ngZh from '@angular/common/locales/zh';
 import ngZhTw from '@angular/common/locales/zh-Hant';
 import { Injectable } from '@angular/core';
 import {
-  AlainI18NService,
   DelonLocaleService,
   en_US as delonEnUS,
   SettingsService,
   zh_CN as delonZhCn,
   zh_TW as delonZhTw,
+  _HttpClient,
+  AlainI18nBaseService
 } from '@delon/theme';
-import { TranslateService } from '@ngx-translate/core';
+import { AlainConfigService } from '@delon/util/config';
 import { enUS as dfEn, zhCN as dfZhCn, zhTW as dfZhTw } from 'date-fns/locale';
 import { NzSafeAny } from 'ng-zorro-antd/core/types';
 import { en_US as zorroEnUS, NzI18nService, zh_CN as zorroZhCN, zh_TW as zorroZhTW } from 'ng-zorro-antd/i18n';
-import { BehaviorSubject, Observable } from 'rxjs';
-import { filter } from 'rxjs/operators';
+import { Observable } from 'rxjs';
 
-interface LangData {
+interface LangConfigData {
   abbr: string;
   text: string;
   ng: NzSafeAny;
@@ -30,14 +29,14 @@ interface LangData {
 }
 
 const DEFAULT = 'zh-CN';
-const LANGS: { [key: string]: LangData } = {
+const LANGS: { [key: string]: LangConfigData } = {
   'zh-CN': {
     text: '简体中文',
     ng: ngZh,
     zorro: zorroZhCN,
     date: dfZhCn,
     delon: delonZhCn,
-    abbr: '🇨🇳',
+    abbr: '🇨🇳'
   },
   'zh-TW': {
     text: '繁体中文',
@@ -45,7 +44,7 @@ const LANGS: { [key: string]: LangData } = {
     zorro: zorroZhTW,
     date: dfZhTw,
     delon: delonZhTw,
-    abbr: '🇭🇰',
+    abbr: '🇭🇰'
   },
   'en-US': {
     text: 'English',
@@ -53,37 +52,30 @@ const LANGS: { [key: string]: LangData } = {
     zorro: zorroEnUS,
     date: dfEn,
     delon: delonEnUS,
-    abbr: '🇬🇧',
-  },
+    abbr: '🇬🇧'
+  }
 };
 
-@Injectable({ providedIn: 'root' })
-export class I18NService implements AlainI18NService {
-  private _default = DEFAULT;
-  private change$ = new BehaviorSubject<string | null>(null);
-
-  private _langs = Object.keys(LANGS).map((code) => {
+@Injectable({providedIn: 'root'})
+export class I18NService extends AlainI18nBaseService {
+  protected override _defaultLang = DEFAULT;
+  private _langs = Object.keys(LANGS).map(code => {
     const item = LANGS[code];
-    return { code, text: item.text, abbr: item.abbr };
+    return {code, text: item.text, abbr: item.abbr};
   });
 
   constructor(
+    private http: _HttpClient,
     private settings: SettingsService,
     private nzI18nService: NzI18nService,
     private delonLocaleService: DelonLocaleService,
-    private translate: TranslateService,
     private platform: Platform,
+    cogSrv: AlainConfigService
   ) {
-    // `@ngx-translate/core` 预先知道支持哪些语言
-    const lans = this._langs.map((item) => item.code);
-    translate.addLangs(lans);
+    super(cogSrv);
 
-    const defaultLan = this.getDefaultLang();
-    if (lans.includes(defaultLan)) {
-      this._default = defaultLan;
-    }
-
-    this.updateLangData(this._default);
+    const defaultLang = this.getDefaultLang();
+    this._defaultLang = this._langs.findIndex(w => w.code === defaultLang) === -1 ? DEFAULT : defaultLang;
   }
 
   private getDefaultLang(): string {
@@ -93,43 +85,31 @@ export class I18NService implements AlainI18NService {
     if (this.settings.layout.lang) {
       return this.settings.layout.lang;
     }
-    return (navigator.languages ? navigator.languages[0] : null) || navigator.language;
+    let res = (navigator.languages ? navigator.languages[0] : null) || navigator.language;
+    const arr = res.split('-');
+    return arr.length <= 1 ? res : `${arr[0]}-${arr[1].toUpperCase()}`;
   }
 
-  private updateLangData(lang: string): void {
+  loadLangData(lang: string): Observable<NzSafeAny> {
+    return this.http.get(`assets/tmp/i18n/${lang}.json`);
+  }
+
+  use(lang: string, data: Record<string, unknown>): void {
+    if (this._currentLang === lang) return;
+
+    this._data = this.flatData(data, []);
+
     const item = LANGS[lang];
     registerLocaleData(item.ng);
     this.nzI18nService.setLocale(item.zorro);
     this.nzI18nService.setDateLocale(item.date);
     this.delonLocaleService.setLocale(item.delon);
+    this._currentLang = lang;
+
+    this._change$.next(lang);
   }
 
-  get change(): Observable<string> {
-    return this.change$.asObservable().pipe(filter((w) => w != null)) as Observable<string>;
-  }
-
-  use(lang: string): void {
-    lang = lang || this.translate.getDefaultLang();
-    if (this.currentLang === lang) {
-      return;
-    }
-    this.updateLangData(lang);
-    this.translate.use(lang).subscribe(() => this.change$.next(lang));
-  }
-  /** 获取语言列表 */
   getLangs(): Array<{ code: string; text: string; abbr: string }> {
     return this._langs;
-  }
-  /** 翻译 */
-  fanyi(key: string, interpolateParams?: {}): any {
-    return this.translate.instant(key, interpolateParams);
-  }
-  /** 默认语言 */
-  get defaultLang(): string {
-    return this._default;
-  }
-  /** 当前语言 */
-  get currentLang(): string {
-    return this.translate.currentLang || this.translate.getDefaultLang() || this._default;
   }
 }
