@@ -8,10 +8,9 @@ import {
   HostBinding,
   Input,
   OnDestroy,
-  Output,
+  Output
 } from '@angular/core';
-import { BehaviorSubject } from 'rxjs';
-import { debounceTime, distinctUntilChanged, tap } from 'rxjs/operators';
+import { BehaviorSubject, debounceTime, distinctUntilChanged, tap } from 'rxjs';
 
 @Component({
   selector: 'header-search',
@@ -31,14 +30,15 @@ import { debounceTime, distinctUntilChanged, tap } from 'rxjs/operators';
         (input)="search($event)"
         (focus)="qFocus()"
         (blur)="qBlur()"
-        [attr.placeholder]="'menu.search.placeholder' | translate"
+        hotkey="F1"
+        [attr.placeholder]="'menu.search.placeholder' | i18n"
       />
     </nz-input-group>
     <nz-autocomplete nzBackfill #auto>
       <nz-auto-option *ngFor="let i of options" [nzValue]="i">{{ i }}</nz-auto-option>
     </nz-autocomplete>
   `,
-  changeDetection: ChangeDetectionStrategy.OnPush,
+  changeDetection: ChangeDetectionStrategy.OnPush
 })
 export class HeaderSearchComponent implements AfterViewInit, OnDestroy {
   q = '';
@@ -63,9 +63,12 @@ export class HeaderSearchComponent implements AfterViewInit, OnDestroy {
       setTimeout(() => this.qIpt!.focus());
     }
   }
-  @Output() toggleChangeChange = new EventEmitter<boolean>();
+  @Output() readonly toggleChangeChange = new EventEmitter<boolean>();
 
-  constructor(private el: ElementRef<HTMLElement>, private cdr: ChangeDetectorRef) {}
+  constructor(
+    private el: ElementRef<HTMLElement>,
+    private cdr: ChangeDetectorRef
+  ) {}
 
   ngAfterViewInit(): void {
     this.qIpt = this.el.nativeElement.querySelector('.ant-input') as HTMLInputElement;
@@ -76,10 +79,10 @@ export class HeaderSearchComponent implements AfterViewInit, OnDestroy {
         tap({
           complete: () => {
             this.loading = true;
-          },
-        }),
+          }
+        })
       )
-      .subscribe((value) => {
+      .subscribe(value => {
         this.options = value ? [value, value + value, value + value + value] : [];
         this.loading = false;
         this.cdr.detectChanges();
