@@ -1,10 +1,10 @@
 import { Component, OnInit, Renderer2, ElementRef } from '@angular/core';
-import { Router, NavigationEnd } from '@angular/router';
+import { Router, NavigationEnd, RouteConfigLoadStart, NavigationError } from '@angular/router';
 import { filter } from 'rxjs/operators';
 import { TitleService, VERSION as VERSION_ALAIN } from '@delon/theme';
 import { NzModalService } from 'ng-zorro-antd/modal';
 import { VERSION as VERSION_ZORRO } from 'ng-zorro-antd/version';
-import { TranslateService } from "@ngx-translate/core";
+import { environment } from "@env/environment";
 
 @Component({
   selector: 'app-root',
@@ -25,6 +25,27 @@ export class AppComponent implements OnInit {
   }
 
   ngOnInit() {
+    let configLoad = false;
+    this.router.events.subscribe(ev => {
+      if (ev instanceof RouteConfigLoadStart) {
+        configLoad = true;
+      }
+      if (configLoad && ev instanceof NavigationError) {
+        this.modalSrv.confirm({
+          nzTitle: `提醒`,
+          nzContent: environment.production ? `应用可能已发布新版本，请点击刷新才能生效。` : `无法加载路由：${ev.url}`,
+          nzCancelDisabled: false,
+          nzOkText: '刷新',
+          nzCancelText: '忽略',
+          nzOnOk: () => location.reload()
+        });
+      }
+      if (ev instanceof NavigationEnd) {
+        this.titleSrv.setTitle();
+        this.modalSrv.closeAll();
+      }
+    });
+
     this.router.events.pipe(filter(evt => evt instanceof NavigationEnd)).subscribe(() => {
       this.titleSrv.setTitle();
       this.modalSrv.closeAll();
