@@ -1,5 +1,5 @@
 // tslint:disable: no-duplicate-imports
-import { NgModule, LOCALE_ID, APP_INITIALIZER } from '@angular/core';
+import { NgModule, LOCALE_ID, APP_INITIALIZER, Type } from '@angular/core';
 import { HttpClient, HttpClientModule } from '@angular/common/http';
 import { BrowserModule } from '@angular/platform-browser';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
@@ -7,53 +7,48 @@ import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 // #region default language
 // Reference: https://ng-alain.com/docs/i18n
 import { default as ngLang } from '@angular/common/locales/zh';
-import { NZ_I18N, en_US as zorroLang } from 'ng-zorro-antd/i18n';
+import { NZ_DATE_LOCALE, NZ_I18N, zh_CN as zorroLang } from 'ng-zorro-antd/i18n';
 import { DELON_LOCALE, en_US as delonLang } from '@delon/theme';
+import { zhCN as dateLang } from 'date-fns/locale';
 
 const LANG = {
-  abbr: 'en',
+  abbr: 'zh',
   ng: ngLang,
   zorro: zorroLang,
   delon: delonLang,
+  date: dateLang
 };
 // register angular
 import { registerLocaleData } from '@angular/common';
+
 registerLocaleData(LANG.ng, LANG.abbr);
 
 const LANG_PROVIDES = [
   {provide: LOCALE_ID, useValue: LANG.abbr},
   {provide: NZ_I18N, useValue: LANG.zorro},
-  {provide: DELON_LOCALE, useValue: LANG.delon}
+  {provide: DELON_LOCALE, useValue: LANG.delon},
+  {provide: NZ_DATE_LOCALE, useValue: LANG.date},
 ];
 // #endregion
 
 // #region i18n services
 import { I18NService } from '@core';
 import { ALAIN_I18N_TOKEN } from '@delon/theme';
-import { TranslateLoader, TranslateModule } from '@ngx-translate/core';
-import { TranslateHttpLoader } from '@ngx-translate/http-loader';
 
-// 加载i18n语言文件
-export function I18nHttpLoaderFactory(http: HttpClient): TranslateHttpLoader {
-  return new TranslateHttpLoader(http, `assets/i18n/`, '.json');
-}
+const I18NSERVICE_PROVIDES = [{provide: ALAIN_I18N_TOKEN, useClass: I18NService, multi: false}];
 
-const I18NSERVICE_MODULES = [
-  TranslateModule.forRoot({
-    loader: {
-      provide: TranslateLoader,
-      useFactory: I18nHttpLoaderFactory,
-      deps: [HttpClient],
-    },
-  }),
-];
+// #endregion
 
-const I18NSERVICE_PROVIDES = [{ provide: ALAIN_I18N_TOKEN, useClass: I18NService, multi: false }];
+// #region global third module
+
+import { BidiModule } from '@angular/cdk/bidi';
+
+const GLOBAL_THIRD_MODULES: Array<Type<any>> = [BidiModule];
 
 // #endregion
 
 // #region JSON Schema form (using @delon/form)
-import { JsonSchemaModule } from '@shared/json-schema/json-schema.module';
+import { JsonSchemaModule } from '@shared';
 
 const FORM_MODULES = [JsonSchemaModule];
 // #endregion
@@ -62,23 +57,15 @@ const FORM_MODULES = [JsonSchemaModule];
 // #region Http Interceptors
 import { HTTP_INTERCEPTORS } from '@angular/common/http';
 import { DefaultInterceptor } from '@core';
-
-import { CoreModule } from '@core/core.module';
-import { SharedModule } from '@shared';
-import { AppComponent } from './app.component';
-import { RoutesModule } from './routes/routes.module';
-import { LayoutModule } from './layout/layout.module';
-import { GlobalConfigModule } from "./global-config.module";
 import { JWTInterceptor } from "@delon/auth";
 
 const INTERCEPTOR_PROVIDES = [
   {provide: HTTP_INTERCEPTORS, useClass: JWTInterceptor, multi: true},
   {provide: HTTP_INTERCEPTORS, useClass: DefaultInterceptor, multi: true}
 ];
-// #endregion
 
-// #region global third module
-const GLOBAL_THIRD_MODULES = [];
+
+
 // #endregion
 
 // #region Startup Service
@@ -101,6 +88,13 @@ const APPINIT_PROVIDES = [
 // #endregion
 
 
+import { CoreModule } from '@core/core.module';
+import { SharedModule } from '@shared';
+import { AppComponent } from './app.component';
+import { RoutesModule } from './routes/routes.module';
+import { LayoutModule } from './layout/layout.module';
+import { GlobalConfigModule } from "./global-config.module";
+
 @NgModule({
   declarations: [
     AppComponent
@@ -114,7 +108,6 @@ const APPINIT_PROVIDES = [
     SharedModule,
     LayoutModule,
     RoutesModule,
-    I18NSERVICE_MODULES,
     ...FORM_MODULES,
     ...GLOBAL_THIRD_MODULES
   ],
