@@ -1,6 +1,5 @@
-import { Injectable, Injector, Inject } from '@angular/core';
+import { Injectable, Inject } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
-import { catchError } from 'rxjs/operators';
 import { ALAIN_I18N_TOKEN, MenuService, SettingsService, TitleService } from '@delon/theme';
 import { DA_SERVICE_TOKEN, ITokenService } from '@delon/auth';
 import { ACLService } from '@delon/acl';
@@ -8,10 +7,11 @@ import { ACLService } from '@delon/acl';
 import { NzIconService } from 'ng-zorro-antd/icon';
 import { ICONS_AUTO } from '../../../style-icons-auto';
 import { ICONS } from '../../../style-icons';
-import { zip } from "rxjs";
+import { Observable, zip, catchError, map } from "rxjs";
 import { EnumService } from "@services";
-import { TranslateService } from "@ngx-translate/core";
 import { I18NService } from "@core/i18n/i18n.service";
+import { NzSafeAny } from "ng-zorro-antd/core/types";
+import { Router } from "@angular/router";
 
 /**
  * Used for application startup
@@ -22,88 +22,64 @@ export class StartupService {
   constructor(
     iconSrv: NzIconService,
     private menuService: MenuService,
-    private enumService:EnumService,
+    private enumService: EnumService,
     private settingService: SettingsService,
     private aclService: ACLService,
     private titleService: TitleService,
     @Inject(DA_SERVICE_TOKEN) private tokenService: ITokenService,
-    private translate: TranslateService,
     @Inject(ALAIN_I18N_TOKEN) private i18n: I18NService,
+    private router: Router,
     private httpClient: HttpClient,
   ) {
     iconSrv.addIcon(...ICONS_AUTO, ...ICONS);
   }
 
-  private viaHttp(resolve: any, reject: any) {
-    zip(
-      this.httpClient.get('assets/tmp/app-data.json')
-    ).pipe(
-      catchError(([appData]) => {
-        resolve(null);
-        return [appData];
-      })
-    ).subscribe(([appData]) => {
-        // Application data
-        const res: any = appData;
-        // Application information: including site name, description, year
-        this.settingService.setApp(res.app);
-        // User information: including name, avatar, email address
-        this.settingService.setUser(res.user);
-        // ACL: Set the permissions to full, https://ng-alain.com/acl/getting-started
-        this.aclService.setFull(true);
-        // Menu data, https://ng-alain.com/theme/menu
-        this.menuService.add(res.menu);
-        // Can be set page suffix title, https://ng-alain.com/theme/title
-        this.titleService.prefix = res.app.name;
-      },
-      () => {
-      },
-      () => {
-        resolve(null);
-      });
-  }
-
-  load(): Promise<any> {
+  load(): Observable<void> {
     // only works with promises
     // https://github.com/angular/angular/issues/15088
-    return new Promise((resolve, reject) => {
-      // http
-      // this.viaHttp(resolve, reject);
-      // mock：请勿在生产环境中这么使用，viaMock 单纯只是为了模拟一些数据使脚手架一开始能正常运行
-      // this.viaMock(resolve, reject);
 
+    const defaultLang = this.i18n.defaultLang;
+    let user = this.settingService.user;
+    return zip(this.i18n.loadLangData(defaultLang), this.httpClient.get(`api/users/${user.id}/menus/tree`)).pipe(
+      // 接收其他拦截器后产生的异常消息
+      catchError(res => {
+        console.warn(`StartupService.load: Network request failed`, res);
+        setTimeout(() => this.router.navigateByUrl(`/exception/500`));
+        return [];
+      }),
+      map(([langData, menuTree]: [Record<string, string>, NzSafeAny]) => {
+        // setting language data
+        this.i18n.use(defaultLang, langData);
 
-      const app: any = {
-        name: `Netcool.Admin`,
-        description: `Netcool.Admin front-end.`
-      };
-      this.translate.setDefaultLang(this.i18n.defaultLang);
-      this.settingService.setApp(app);
-      this.titleService.prefix = app.name;
+        // 应用信息：包括站点名、描述、年份
+        const app: any = {
+          name: `Netcool.Admin`,
+          description: `Netcool.Admin front-end.`
+        };
+        this.settingService.setApp(app);
 
-      this.enumService.loadEnums();
+        // 设置页面标题的后缀
+        this.titleService.default = '';
+        this.titleService.prefix = app.name;
+        //this.settingService.setApp(appData.app);
 
-      let user = this.settingService.user;
-      this.aclService.setFull(false)
-      this.aclService.setAbility(user.permissionCodes);
+        this.enumService.loadEnums();
 
-      this.httpClient.get(`api/users/${user.id}/menus/tree`)
-        .subscribe((tree: any) => {
-            let rootMenu = {
-              text: '导航',
-              group: true,
-              children: []
-            };
-            this.appendChildren(rootMenu, tree);
-            this.menuService.add([rootMenu]);
-          },
-          error => {
-            console.error(error);
-            resolve(null);
-          }, () => {
-            resolve(null);
-          });
-    });
+        // 用户信息：包括姓名、头像、邮箱地址
+
+        this.aclService.setFull(false)
+        this.aclService.setAbility(user.permissionCodes);
+        // 初始化菜单
+        let rootMenu = {
+          text: '导航',
+          group: true,
+          children: []
+        };
+        this.appendChildren(rootMenu, menuTree);
+        this.menuService.add([rootMenu]);
+
+      })
+    );
   }
 
   private appendChildren(parent, tree): void {
