@@ -1,33 +1,29 @@
 import { ChangeDetectionStrategy, Component, HostListener } from '@angular/core';
-import * as screenfull from 'screenfull';
+import screenfull from 'screenfull';
 
 @Component({
   selector: 'header-fullscreen',
   template: `
     <i nz-icon [nzType]="status ? 'fullscreen-exit' : 'fullscreen'"></i>
-    {{ (status ? 'menu.fullscreen.exit' : 'menu.fullscreen') | translate }}
+    {{ (status ? 'menu.fullscreen.exit' : 'menu.fullscreen') | i18n }}
   `,
-  // tslint:disable-next-line: no-host-metadata-property
   host: {
-    '[class.d-block]': 'true',
+    '[class.flex-1]': 'true'
   },
-  changeDetection: ChangeDetectionStrategy.OnPush,
+  changeDetection: ChangeDetectionStrategy.OnPush
 })
 export class HeaderFullScreenComponent {
   status = false;
-  private get sf(): screenfull.Screenfull {
-    return screenfull as screenfull.Screenfull;
-  }
 
   @HostListener('window:resize')
   _resize(): void {
-    this.status = this.sf.isFullscreen;
+    this.status = screenfull.isFullscreen;
   }
 
   @HostListener('click')
   _click(): void {
-    if (this.sf.isEnabled) {
-      this.sf.toggle();
+    if (screenfull.isEnabled) {
+      screenfull.toggle();
     }
   }
 }
