@@ -10,14 +10,16 @@ import { UserLoginComponent } from './passport/login/login.component';
 // single pages
 import { CallbackComponent } from './callback/callback.component';
 import { UserLockComponent } from './passport/lock/lock.component';
-import { SimpleGuard } from "@delon/auth";
+import { authSimpleCanActivate, authSimpleCanActivateChild } from '@delon/auth';
 import { LayoutBasicComponent } from "../layout/basic/basic.component";
+import { PreloadOptionalModules } from '@delon/theme';
 
 const routes: Routes = [
   {
     path: '',
     component: LayoutBasicComponent,
-    canActivate: [SimpleGuard],
+    canActivate: [ authSimpleCanActivate],
+    canActivateChild: [authSimpleCanActivateChild],
     children: [
       {path: '', redirectTo: 'dashboard', pathMatch: 'full'},
       {path: 'dashboard', component: DashboardComponent, data: {title: '仪表盘', titleI18n: 'dashboard'}},
@@ -49,13 +51,16 @@ const routes: Routes = [
 ];
 
 @NgModule({
+  providers: [PreloadOptionalModules],
   imports: [
     RouterModule.forRoot(
       routes, {
         useHash: environment.useHash,
         // NOTICE: If you use `reuse-tab` component and turn on keepingScroll you can set to `disabled`
         // Pls refer to https://ng-alain.com/components/reuse-tab
-        scrollPositionRestoration: 'top'
+        scrollPositionRestoration: 'top',
+        preloadingStrategy: PreloadOptionalModules,
+        bindToComponentInputs: true
       }
     )],
   exports: [RouterModule],
